A changelog markdown template to document project progress

A changelog markdown template to document project progress

Usage

init_changelog_md(path = getwd(), confirm = TRUE)

Arguments

path

Path where markdown file is created

confirm

Confirm path; logical input.

Value

Return markdown file in specified or current working directory

Examples

init_changelog_md()
#> You current working directory will be:
#> /home/runner/work/peacock/peacock/docs/referenceDo you wish to create a Change log MD here? (y/yes to confirm): 
#> Change log MD initialization canceled.